Remove snippets.
I found the snippets to be distracting, especially because the
suggestions appear in positions  where the resulting code is invalid.
Snippets should belong in separate vscode packages with the category
"snippets". That is how it's done for TypeScript:
https://marketplace.visualstudio.com/search?term=typescript&target=VSCode&category=Snippets&sortBy=Relevance

Note that a snippet plugin can be located in the same github repo as the
syntax.